requirements.txt
scikit-learn===0.20.0
pandas===1.0.3
numpy===1.18.4
scipy===1.4.1
matplotlib
keras
tensorflow
pillow
selenium
requests